src/demand/app/requirements.txt
django>=3.2,<4.0
djangorestframework
django_filter
drf-spectacular
psycopg2-binary
django-autoslug
django-cors-headers
django.dotenv
gunicorn
djangorestframework-simplejwt
pika
